Added support for retrieving node errors to further diagnose a failed cook. Also improved handling of partial cook failures when only some objects failed to cook but the asset overall state is still valid.
API Changes:
- HAPI_STATE_READY_WITH_ERRORS has been renamed to HAPI_STATE_READY_WITH_FATAL_ERRORS.
- A new state called HAPI_STATE_READY_WITH_COOK_ERRORS has been added to distinguish between just some objects failing and a truly fatal error like can't find an expected node.
- Added new enum called HAPI_CookErrorSearchMode and corresponding setting HAPI_CookOptions::cookErrorSearchMode that will determine how verbose the cook errors are. The more verbose the more expensive a cook error becomes, hence the need for a setting.

+++ Asset.C
@@ -382,7 +382,7 @@ Asset::Asset(
     if(assetNamesSH.size())
     {
         bool foundAsset = false;
-        for(int i = 0; i < assetNamesSH.size(); i++)
+        for(int i = 0; i < (int) assetNamesSH.size(); i++)
         {
             if(Util::getString(assetNamesSH[i]) == assetName)
             {
